refactor: code 컨벤션 맞춰서 수정 및 파일 경로 변경 #182
