+++ src/app/concert/form/[id]/_shared/validators/application-form.ts
@@ -1,5 +1,5 @@
 import { FormData } from '@/app/concert/form/[id]/_shared/components/form-input/form-input.type';
-import { ValidationConstants } from '@/shared/types/ValidationConstants';
+import { VALIDATION_CONSTANTS } from '@/shared/constants/validation';
 
 export type ValidationResult =
   | { valid: true }
@@ -22,7 +22,7 @@ export const validateApplicationForm = (
     if (
       !data.requestCount ||
       Number(data.requestCount) <
-        ValidationConstants.ApplicationForm.APPLICATION_FORM_MIN_REQUEST_COUNT
+        VALIDATION_CONSTANTS.ApplicationForm.APPLICATION_FORM_MIN_REQUEST_COUNT
     ) {
       return {
         valid: false,
@@ -34,12 +34,12 @@ export const validateApplicationForm = (
     if (
       data.requirement &&
       data.requirement.length >
-        ValidationConstants.ApplicationForm.REQUIREMENT_MAX_LENGTH
+        VALIDATION_CONSTANTS.ApplicationForm.REQUIREMENT_MAX_LENGTH
     ) {
       return {
         valid: false,
         tabId: numericTabId,
-        message: `요청사항은 최대 ${ValidationConstants.ApplicationForm.REQUIREMENT_MAX_LENGTH}자까지 입력 가능합니다.`,
+        message: `요청사항은 최대 ${VALIDATION_CONSTANTS.ApplicationForm.REQUIREMENT_MAX_LENGTH}자까지 입력 가능합니다.`,
       };
     }
   }
